all module have been splitted except a small bug in forum page
