+++ src/poc-data/extracted_course_works.csv
@@ -28,3 +28,6 @@ TEST,Final Project: Creative Artwork,Not Found,2024-04-10,50
 4C03,Assignment 2: Essay,Not Found,2024-12-07,25
 test2,Assignment 1: Deconstructed Essay,Not Found,2024-10-16,25
 test2,Assignment 2: Essay,Not Found,2024-12-07,25
+4A03,Assignment 1: Deconstructed Essay,Not Found,2024-10-16,25
+4A03,Test 2,Not Found,2024-11-27,25
+4A03,Assignment 2: Essay,Not Found,2024-12-07,25
