+++ src/poc-data/mock_course_info.csv
@@ -18,3 +18,4 @@ TEST,TEST-syllabus.pdf,Not Found,Not Found,Not Found,Not Found,Not Found,Not Fou
    Available at McMaster’s Campus Store (CAD) as a print or e-text. 
    ISBN-13: 9780190177928, eISBN-13: 0190177926",Asynchronous Online,Not Found,Not Found,"This course introduces students to the field of Anthropology and its sub-disciplines through discussing various anthropological perspectives on sex, food, illness, and death from archaeological, biological, sociocultural, and linguistic viewpoints. Students will develop a better understanding of how anthropological approaches help in understanding the human condition past and present.",Not Found,Not Found
 test2,test2-syllabus.pdf,Beatrice Fletcher,fletchba@mcmaster.ca,Tuesday 2-3pm or by appointment,"Introduction to Anthropology - Sex, Food, and Death (Custom edition for McMaster University, Oxford University Press, 2023) ISBN-13: 9780190177928, eISBN-13: 0190177926 (Note: Textbook updated as of September 2023)",Asynchronous Online,Not Found,Not Found,"This course introduces students to the field of Anthropology and its sub-disciplines through exploring topics such as sex, gender, food, illness, and death from various anthropological perspectives. Students will develop critical thinking and analytical skills to understand the human condition past and present.",Not Found,Not Found
+4A03,4A03-syllabus.pdf,Beatrice Fletcher,fletchba@mcmaster.ca,Tuesday 2-3pm or by appointment,"Introduction to Anthropology - Sex, Food and Death, Custom edition for McMaster University, Oxford University Press, 2023, ISBN-13: 9780190177928, eISBN-13: 0190177926 (Note: Text has been updated, effective September 2023)",Asynchronous Online,Not Found,Not Found,"This course introduces students to the field of Anthropology and its sub-disciplines through the discussion of big questions anthropologists investigate, focusing on sex, gender, food, illness, and death from various perspectives. It aims to develop an understanding of the human condition past and present.",Not Found,Not Found
